Conflicts:
	scipy/interpolate/rbf.py

+++ scipy/io/arff/tests/test_arffread.py
@@ -39,7 +39,7 @@ expected_types = ['numeric', 'numeric', 'numeric', 'numeric', 'nominal']
 
 missing = pjoin(data_path, 'missing.arff')
 expect_missing_raw = np.array([[1, 5], [2, 4], [np.nan, np.nan]])
-expect_missing = np.empty(3, [('yop', np.float), ('yap', np.float)])
+expect_missing = np.empty(3, [('yop', float), ('yap', float)])
 expect_missing['yop'] = expect_missing_raw[:, 0]
 expect_missing['yap'] = expect_missing_raw[:, 1]
